requirements.txt
alive_progress==3.1.0
docx==0.2.4
docxtpl==0.16.6
Jinja2==3.1.2
Pillow==9.5.0
python_docx==0.8.11
requests==2.28.1
